+++ stm32/libraries/USBDevice/src/USBDevice.cpp
@@ -1,17 +1,8 @@
-#ifndef _TINYUSB_H_
-#define _TINYUSB_H_
-
-#include "tusb_option.h"
-#include "tusb.h"
-
-#ifdef TUSB_OPT_DEVICE_ENABLED
-#endif
+#include "USBDevice.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-
 
 /* +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
 #if defined(TINYUSB_BOARD__STM32F072B_DISCO) // BOARD
@@ -40,7 +31,7 @@ void board_init(void) {
 }
 
 /* +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
-#elif defined(TINYUSB_BOARD__KLST_TINY)
+#elif defined(KLST_BOARD_KLST_TINY)
 /* +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
 
 //--------------------------------------------------------------------+
@@ -49,15 +40,6 @@ void board_init(void) {
 void OTG_FS_IRQHandler(void) {
     // @note(always make sure to spec the correct callback. i.e STM32F0 callback is `USB_IRQHandler`)
     tud_int_handler(0);
-}
-
-static inline void board_vbus_sense_init(void) {
-    /* Deactivate VBUS Sensing B */
-    USB_OTG_FS->GCCFG &= ~USB_OTG_GCCFG_VBDEN;
-
-    /* B-peripheral session valid override enable */
-    USB_OTG_FS->GOTGCTL |= USB_OTG_GOTGCTL_BVALOEN;
-    USB_OTG_FS->GOTGCTL |= USB_OTG_GOTGCTL_BVALOVAL;
 }
 
 void board_init(void) {
@@ -90,7 +72,67 @@ void board_init(void) {
     // Enable USB OTG clock
     __HAL_RCC_USB_OTG_FS_CLK_ENABLE();
 
-    board_vbus_sense_init();
+    /* Deactivate VBUS Sensing B */
+    USB_OTG_FS->GCCFG &= ~USB_OTG_GCCFG_VBDEN;
+
+    /* B-peripheral session valid override enable */
+    USB_OTG_FS->GOTGCTL |= USB_OTG_GOTGCTL_BVALOEN;
+    USB_OTG_FS->GOTGCTL |= USB_OTG_GOTGCTL_BVALOVAL;}
+
+/* +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
+#elif defined(KLST_BOARD_KLST_CORE)
+/* +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
+
+// void OTG_HS_IRQHandler(void) {
+//     tud_int_handler(1);
+// }
+
+void OTG_FS_IRQHandler(void) {
+    tud_int_handler(0);
+}
+
+void board_init(void) {
+    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    RCC_PeriphCLKInitTypeDef PeriphClkInitStruct = {0};
+
+    PeriphClkInitStruct.PeriphClockSelection = RCC_PERIPHCLK_USB;
+    PeriphClkInitStruct.UsbClockSelection = RCC_USBCLKSOURCE_HSI48;
+    if (HAL_RCCEx_PeriphCLKConfig(&PeriphClkInitStruct) != HAL_OK)
+    {
+      Error_Handler();
+    }
+
+    /** Enable USB Voltage detector */
+    HAL_PWREx_EnableUSBVoltageDetector();
+
+    __HAL_RCC_GPIOA_CLK_ENABLE();
+    /* Configure DM DP Pins */
+    GPIO_InitStruct.Pin = GPIO_PIN_11 | GPIO_PIN_12;
+    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
+    GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
+    GPIO_InitStruct.Pull = GPIO_NOPULL;
+    GPIO_InitStruct.Alternate = GPIO_AF10_OTG2_HS;
+    HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+    /* This for ID line debug */
+    GPIO_InitStruct.Pin = GPIO_PIN_10;
+    GPIO_InitStruct.Mode = GPIO_MODE_AF_OD;
+    GPIO_InitStruct.Pull = GPIO_PULLUP;
+    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_HIGH;
+    GPIO_InitStruct.Alternate = GPIO_AF10_OTG2_HS;
+    HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+    // https://community.st.com/s/question/0D50X00009XkYZLSA3/stm32h7-nucleo-usb-fs-cdc
+    // TODO: Board init actually works fine without this line.
+    HAL_PWREx_EnableUSBVoltageDetector();
+    __HAL_RCC_USB2_OTG_FS_CLK_ENABLE();
+
+    // Disable VBUS sense (B device) via pin PA9
+    USB_OTG_FS->GCCFG &= ~USB_OTG_GCCFG_VBDEN;
+
+    // B-peripheral session valid override enable
+    USB_OTG_FS->GOTGCTL |= USB_OTG_GOTGCTL_BVALOEN;
+    USB_OTG_FS->GOTGCTL |= USB_OTG_GOTGCTL_BVALOVAL;
 }
 
 /* +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
@@ -105,5 +147,3 @@ void board_init(void) {}
 #ifdef __cplusplus
 }
 #endif
-
-#endif /* _TINYUSB_H_ */